Adding scripts to new folders
