+++ Assets/Scenes/StartMenu.unity
@@ -2130,6 +2130,7 @@ GameObject:
   - component: {fileID: 1353647014}
   - component: {fileID: 1353647013}
   - component: {fileID: 1353647012}
+  - component: {fileID: 1353647015}
   m_Layer: 5
   m_Name: ExitButton
   m_TagString: Untagged
@@ -2252,6 +2253,18 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1353647010}
   m_CullTransparentMesh: 1
+--- !u!114 &1353647015
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1353647010}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 1e2e03e01d6360b47b597e8112f8e492, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &1408729428
 GameObject:
   m_ObjectHideFlags: 0
