--- Assets/Scripts/Skin.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 073649bd44de6a049866d2df3aae146b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
