+++ Assets/Scripts/StartNewsMenu.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4225b9e6f702b824eb337db4eeb0b6e8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
