+++ Assets/Scripts/StartNewsMenu/UpdateManager.cs
